ft_transcendance/backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
django-cors-headers
psycopg2-binary>=2.9